requirements.txt
numpy
opencv-python
paddlepaddle_gpu>=2.2
Pillow==8.4.0
pyyaml
wandb
